feat:connected typescript
